+++ Changelog.md
@@ -2,6 +2,7 @@
 
 - Fixes race condition when displaying image animations (e.g. gifs via sixel).
 - Fixes `NewTerminal` action to also inherit the active configuration file.
+- Fixes restoring cursor position in `RM ?1049`.
 - Adds VT sequence `OSC 7` (set current working directory).
 
 ### 0.1.0 (2012-12-24)